Refactor large UI and notation files into smaller modules
