+++ lib/features/capture/presentation/import_score_screen.dart
@@ -16,6 +16,7 @@ import 'package:note_vision/features/musicXML/musicxml_score_converter.dart';
 import 'package:note_vision/features/musicXML/musicxml_validator_service.dart';
 import 'package:note_vision/core/widgets/score_notation_viewer.dart';
 import 'package:note_vision/features/capture/presentation/capture_screen.dart';
+import 'package:note_vision/features/capture/presentation/widgets/capture_shared_widgets.dart';
 
 class ImportScoreScreen extends StatefulWidget {
   const ImportScoreScreen({super.key});
@@ -306,7 +307,7 @@ class _ImportScoreScreenState extends State<ImportScoreScreen> {
         children: [
           _buildContinueButton(),
           const SizedBox(height: 10),
-          _TappableButton(
+          CaptureTappableButton(
             onPressed: _cancelSelection,
             child: Container(
               width: double.infinity,
@@ -354,7 +355,7 @@ class _ImportScoreScreenState extends State<ImportScoreScreen> {
     required VoidCallback? onPressed,
   }) {
     return Expanded(
-      child: _TappableButton(
+      child: CaptureTappableButton(
         onPressed: onPressed,
         child: Container(
           height: 52,
@@ -396,7 +397,7 @@ class _ImportScoreScreenState extends State<ImportScoreScreen> {
     required VoidCallback? onPressed,
   }) {
     return Expanded(
-      child: _TappableButton(
+      child: CaptureTappableButton(
         onPressed: onPressed,
         child: Container(
           height: 52,
@@ -426,7 +427,7 @@ class _ImportScoreScreenState extends State<ImportScoreScreen> {
   }
 
   Widget _buildContinueButton() {
-    return _TappableButton(
+    return CaptureTappableButton(
       onPressed: _continueToEditor,
       child: Container(
         width: double.infinity,
@@ -474,7 +475,7 @@ class _ImportScoreScreenState extends State<ImportScoreScreen> {
           height: 60,
           child: Row(
             children: [
-              _BottomNavItem(
+              CaptureBottomNavItem(
                 icon: Icons.camera_alt_outlined,
                 label: 'Scan',
                 isSelected: false,
@@ -485,7 +486,7 @@ class _ImportScoreScreenState extends State<ImportScoreScreen> {
                   );
                 },
               ),
-              _BottomNavItem(
+              CaptureBottomNavItem(
                 icon: Icons.upload_file_outlined,
                 label: 'Import',
                 isSelected: true,
@@ -736,90 +737,6 @@ class _ImportScoreScreenState extends State<ImportScoreScreen> {
           'quarter',
       voice: (symbol['voice'] as num?)?.toInt(),
       staff: (symbol['staff'] as num?)?.toInt(),
-    );
-  }
-}
-
-class _BottomNavItem extends StatelessWidget {
-  final IconData icon;
-  final String label;
-  final bool isSelected;
-  final VoidCallback onTap;
-
-  const _BottomNavItem({
-    required this.icon,
-    required this.label,
-    required this.isSelected,
-    required this.onTap,
-  });
-
-  static const _accent = Color(0xFFD4A96A);
-  static const _textSecondary = Color(0xFF8A8A8A);
-
-  @override
-  Widget build(BuildContext context) {
-    return Expanded(
-      child: GestureDetector(
-        onTap: onTap,
-        behavior: HitTestBehavior.opaque,
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(
-              icon,
-              size: 22,
-              color: isSelected ? _accent : _textSecondary,
-            ),
-            const SizedBox(height: 4),
-            Text(
-              label,
-              style: TextStyle(
-                fontSize: 11,
-                color: isSelected ? _accent : _textSecondary,
-                fontWeight: isSelected ? FontWeight.w600 : FontWeight.w400,
-                letterSpacing: 0.3,
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _TappableButton extends StatefulWidget {
-  final Widget child;
-  final VoidCallback? onPressed;
-
-  const _TappableButton({required this.child, required this.onPressed});
-
-  @override
-  State<_TappableButton> createState() => _TappableButtonState();
-}
-
-class _TappableButtonState extends State<_TappableButton> {
-  bool _pressed = false;
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTapDown: widget.onPressed == null ? null : (_) => setState(() => _pressed = true),
-      onTapCancel: widget.onPressed == null ? null : () => setState(() => _pressed = false),
-      onTap: widget.onPressed == null
-          ? null
-          : () {
-              setState(() => _pressed = false);
-              widget.onPressed!();
-            },
-      child: AnimatedScale(
-        scale: _pressed ? 0.97 : 1.0,
-        duration: const Duration(milliseconds: 100),
-        child: AnimatedOpacity(
-          opacity: _pressed ? 0.85 : 1.0,
-          duration: const Duration(milliseconds: 100),
-          child: widget.child,
-        ),
-      ),
     );
   }
 }
