+++ lib/features/capture/presentation/widgets/capture_shared_widgets.dart
@@ -1,0 +1,90 @@
+import 'package:flutter/material.dart';
+
+class CaptureBottomNavItem extends StatelessWidget {
+  const CaptureBottomNavItem({
+    super.key,
+    required this.icon,
+    required this.label,
+    required this.isSelected,
+    required this.onTap,
+  });
+
+  final IconData icon;
+  final String label;
+  final bool isSelected;
+  final VoidCallback onTap;
+
+  static const _accent = Color(0xFFD4A96A);
+  static const _textSecondary = Color(0xFF8A8A8A);
+
+  @override
+  Widget build(BuildContext context) {
+    return Expanded(
+      child: GestureDetector(
+        onTap: onTap,
+        behavior: HitTestBehavior.opaque,
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Icon(
+              icon,
+              size: 22,
+              color: isSelected ? _accent : _textSecondary,
+            ),
+            const SizedBox(height: 4),
+            Text(
+              label,
+              style: TextStyle(
+                fontSize: 11,
+                color: isSelected ? _accent : _textSecondary,
+                fontWeight: isSelected ? FontWeight.w600 : FontWeight.w400,
+                letterSpacing: 0.3,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+class CaptureTappableButton extends StatefulWidget {
+  const CaptureTappableButton({
+    super.key,
+    required this.child,
+    required this.onPressed,
+  });
+
+  final Widget child;
+  final VoidCallback? onPressed;
+
+  @override
+  State<CaptureTappableButton> createState() => _CaptureTappableButtonState();
+}
+
+class _CaptureTappableButtonState extends State<CaptureTappableButton> {
+  bool _pressed = false;
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTapDown: widget.onPressed == null ? null : (_) => setState(() => _pressed = true),
+      onTapCancel: widget.onPressed == null ? null : () => setState(() => _pressed = false),
+      onTap: widget.onPressed == null
+          ? null
+          : () {
+              setState(() => _pressed = false);
+              widget.onPressed!();
+            },
+      child: AnimatedScale(
+        scale: _pressed ? 0.97 : 1.0,
+        duration: const Duration(milliseconds: 100),
+        child: AnimatedOpacity(
+          opacity: _pressed ? 0.85 : 1.0,
+          duration: const Duration(milliseconds: 100),
+          child: widget.child,
+        ),
+      ),
+    );
+  }
+}
